🚚 Moved i18n files
