+++ src/i18n/en-US/Dialogs/autoshutdown.ts
