+++ src/i18n/en-US/Dialogs/eulaDialog.ts
