+++ src/i18n/en-US/Dialogs/icon.ts
