+++ src/i18n/en-US/Dialogs/owner.ts
