+++ src/i18n/en-US/World/player.ts
