+++ src/i18n/en-US/World/progress.ts
