+++ src/i18n/en-US/World/property.ts
