+++ src/i18n/ja/Dialogs/autoshutdown.ts
