+++ src/i18n/ja/Dialogs/eulaDialog.ts
