+++ src/i18n/ja/Dialogs/icon.ts
