+++ src/i18n/ja/Dialogs/owner.ts
