+++ src/i18n/ja/Dialogs/welcome.ts
