+++ src/i18n/ja/Other/error.ts
