+++ src/i18n/ja/Other/general.ts
