+++ src/i18n/ja/Other/worldList.ts
@@ -1,0 +1,15 @@
+export const jaWorldList = {
+  allWorld: 'ワールド一覧',
+  addWorld: 'ワールドを追加',
+  systemSetting: 'システム設定',
+  searchWorld: 'ワールドを検索',
+  addSeveralWorld: '各種ワールドの導入',
+  addCustomWorld: '配布ワールドを導入',
+  selectZip: 'ZIPファイルを選択',
+  selectFolder: 'フォルダを選択',
+  addSingleWorld: 'シングルワールドを導入',
+  loadSingleWorld: 'シングルプレイワールドを<br>読み込み中',
+  noSingleWorld: 'シングルプレイワールドは<br>見つかりませんでした',
+  openList: 'ワールド一覧を開く',
+  minimizeList: 'ワールド一覧を最小化',
+};
