+++ src/i18n/ja/World/console.ts
