+++ src/i18n/ja/World/home.ts
