+++ src/i18n/ja/World/player.ts
