+++ src/i18n/ja/World/progress.ts
