+++ src/i18n/ja/World/property.ts
