+++ src/i18n/ja/World/shareWorld.ts
